backend/requirements.txt
openai>=1.52.0
yt-dlp>=2025.1.15
Jinja2>=3.1.4
requests>=2.32.3
python-dotenv>=1.0.1
fastapi>=0.115.0
uvicorn>=0.30.6
reportlab>=4.2.2
markdown>=3.7
beautifulsoup4>=4.12.3
